Split some functions from TieredBlockStoreTestUtils.java to EvictorTestUtils.java
