+++ servers/src/test/java/tachyon/worker/block/TieredBlockStoreTestUtils.java
@@ -17,13 +17,8 @@ package tachyon.worker.block;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 import org.apache.commons.io.FileUtils;
-import org.junit.Assert;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -32,16 +27,12 @@ import com.google.common.io.Files;
 import com.google.common.primitives.Ints;
 
 import tachyon.Constants;
-import tachyon.Pair;
 import tachyon.conf.TachyonConf;
-import tachyon.exception.NotFoundException;
 import tachyon.util.io.BufferUtils;
 import tachyon.util.io.PathUtils;
-import tachyon.worker.block.evictor.EvictionPlan;
 import tachyon.worker.block.evictor.Evictor;
 import tachyon.worker.block.io.BlockWriter;
 import tachyon.worker.block.io.LocalFileBlockWriter;
-import tachyon.worker.block.meta.BlockMeta;
 import tachyon.worker.block.meta.StorageDir;
 import tachyon.worker.block.meta.TempBlockMeta;
 
@@ -159,7 +150,6 @@ public class TieredBlockStoreTestUtils {
     return newTachyonConf(TIER_LEVEL, TIER_ALIAS, dirs, TIER_CAPACITY);
   }
 
-
   /**
    * Cache bytes into StroageDir.
    *
@@ -198,187 +188,4 @@ public class TieredBlockStoreTestUtils {
     }
   }
 
-  /**
-   * Whether the plan can satisfy the requested free bytes to be available, assume all blocks in the
-   * plan are in the same dir.
-   *
-   * @param bytesToBeAvailable the requested bytes to be available
-   * @param plan the eviction plan, should not be null
-   * @param meta the metadata manager
-   * @return true if the request can be satisfied otherwise false
-   * @throws NotFoundException if can not get meta data of a block
-   */
-  public static boolean requestSpaceSatisfied(long bytesToBeAvailable, EvictionPlan plan,
-      BlockMetadataManager meta) throws NotFoundException {
-    Preconditions.checkNotNull(plan);
-
-    List<Long> blockIds = new ArrayList<Long>();
-    blockIds.addAll(plan.toEvict());
-    for (Pair<Long, BlockStoreLocation> move : plan.toMove()) {
-      blockIds.add(move.getFirst());
-    }
-
-    long evictedOrMovedBytes = 0;
-    for (long blockId : blockIds) {
-      evictedOrMovedBytes += meta.getBlockMeta(blockId).getBlockSize();
-    }
-
-    BlockStoreLocation location =
-        meta.getBlockMeta(blockIds.get(0)).getParentDir().toBlockStoreLocation();
-    return (meta.getAvailableBytes(location) + evictedOrMovedBytes) >= bytesToBeAvailable;
-  }
-
-  /**
-   * Whether blocks in the EvictionPlan are in the same StorageDir.
-   *
-   * @param plan the eviction plan
-   * @param meta the meta data manager
-   * @return true if blocks are in the same dir otherwise false
-   * @throws NotFoundException if fail to get meta data of a block
-   */
-  public static boolean blocksInTheSameDir(EvictionPlan plan, BlockMetadataManager meta)
-      throws NotFoundException {
-    Preconditions.checkNotNull(plan);
-
-    StorageDir dir = null;
-    List<Long> blockIds = new ArrayList<Long>();
-    blockIds.addAll(plan.toEvict());
-    for (Pair<Long, BlockStoreLocation> move : plan.toMove()) {
-      blockIds.add(move.getFirst());
-    }
-
-    for (long blockId : blockIds) {
-      StorageDir blockDir = meta.getBlockMeta(blockId).getParentDir();
-      if (dir == null) {
-        dir = blockDir;
-      } else if (dir != blockDir) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  /**
-   * Assume the plan is returned by a non-cascading evictor, check whether it is valid. a cascading
-   * evictor is an evictor that always tries to move from the target tier to the next tier and
-   * recursively move down 1 tier until finally blocks are evicted from the final tier.
-   *
-   * @param bytesToBeAvailable the requested bytes to be available
-   * @param plan the eviction plan, should not be null
-   * @param metaManager the meta data manager
-   * @return true if and only if the plan is not null and both {@link #blocksInTheSameDir} and
-   *         {@link #requestSpaceSatisfied} are true, otherwise false
-   * @throws NotFoundException when fail to get meta data of a block
-   */
-  public static boolean validNonCascadingPlan(long bytesToBeAvailable, EvictionPlan plan,
-      BlockMetadataManager metaManager) throws NotFoundException {
-    Preconditions.checkNotNull(plan);
-    return blocksInTheSameDir(plan, metaManager)
-        && requestSpaceSatisfied(bytesToBeAvailable, plan, metaManager);
-  }
-
-  /**
-   * Checks whether the plan of a cascading evictor is valid.
-   *
-   * A cascading evictor will try to free space by recursively moving blocks to next 1 tier and
-   * evict blocks only in the bottom tier.
-   *
-   * The plan is invalid when the requested space can not be satisfied or lower level of tiers do
-   * not have enough space to hold blocks moved from higher level of tiers.
-   *
-   * @param bytesToBeAvailable requested bytes to be available after eviction
-   * @param plan the eviction plan, should not be empty
-   * @param metaManager the meta data manager
-   * @return true if the above requirements are satisfied, otherwise false.
-   */
-  // TODO: unit test this method
-  public static boolean validCascadingPlan(long bytesToBeAvailable, EvictionPlan plan,
-      BlockMetadataManager metaManager) throws Exception {
-    // reassure the plan is feasible: enough free space to satisfy bytesToBeAvailable, and enough
-    // space in lower tier to move blocks in upper tier there
-
-    // Map from dir to a pair of bytes to be available in this dir and bytes to move into this dir
-    // after the plan taking action
-    Map<StorageDir, Pair<Long, Long>> spaceInfoInDir = new HashMap<StorageDir, Pair<Long, Long>>();
-
-    for (long blockId : plan.toEvict()) {
-      BlockMeta block = metaManager.getBlockMeta(blockId);
-      StorageDir dir = block.getParentDir();
-      if (spaceInfoInDir.containsKey(dir)) {
-        Pair<Long, Long> spaceInfo = spaceInfoInDir.get(dir);
-        spaceInfo.setFirst(spaceInfo.getFirst() + block.getBlockSize());
-      } else {
-        spaceInfoInDir.put(dir, new Pair<Long, Long>(
-            dir.getAvailableBytes() + block.getBlockSize(), 0L));
-      }
-    }
-
-    for (Pair<Long, BlockStoreLocation> move : plan.toMove()) {
-      long blockId = move.getFirst();
-      BlockMeta block = metaManager.getBlockMeta(blockId);
-      long blockSize = block.getBlockSize();
-      StorageDir srcDir = block.getParentDir();
-      StorageDir destDir = metaManager.getDir(move.getSecond());
-
-      if (spaceInfoInDir.containsKey(srcDir)) {
-        Pair<Long, Long> spaceInfo = spaceInfoInDir.get(srcDir);
-        spaceInfo.setFirst(spaceInfo.getFirst() + blockSize);
-      } else {
-        spaceInfoInDir
-            .put(srcDir, new Pair<Long, Long>(srcDir.getAvailableBytes() + blockSize, 0L));
-      }
-
-      if (spaceInfoInDir.containsKey(destDir)) {
-        Pair<Long, Long> spaceInfo = spaceInfoInDir.get(destDir);
-        spaceInfo.setSecond(spaceInfo.getSecond() + blockSize);
-      } else {
-        spaceInfoInDir.put(destDir, new Pair<Long, Long>(destDir.getAvailableBytes(), blockSize));
-      }
-    }
-
-    // the top tier among all tiers where blocks in the plan reside in
-    int topTierAlias = Integer.MAX_VALUE;
-    for (StorageDir dir : spaceInfoInDir.keySet()) {
-      topTierAlias = Math.min(topTierAlias, dir.getParentTier().getTierAlias());
-    }
-    long maxSpace = Long.MIN_VALUE; // maximum bytes to be available in a dir in the top tier
-    for (StorageDir dir : spaceInfoInDir.keySet()) {
-      if (dir.getParentTier().getTierAlias() == topTierAlias) {
-        Pair<Long, Long> space = spaceInfoInDir.get(dir);
-        maxSpace = Math.max(maxSpace, space.getFirst() - space.getSecond());
-      }
-    }
-    if (maxSpace < bytesToBeAvailable) {
-      // plan is invalid because requested space can not be satisfied in the top tier
-      return false;
-    }
-
-    for (StorageDir dir : spaceInfoInDir.keySet()) {
-      Pair<Long, Long> spaceInfo = spaceInfoInDir.get(dir);
-      if (spaceInfo.getFirst() < spaceInfo.getSecond()) {
-        // plan is invalid because there is not enough space in this dir to hold the blocks waiting
-        // to be moved into this dir
-        return false;
-      }
-    }
-
-    return true;
-  }
-
-  /**
-   * Only when plan is not null and at least one of
-   * {@link TieredBlockStoreTestUtils#validCascadingPlan}, {@link #validNonCascadingPlan} is true,
-   * the assertion will be passed, used in unit test.
-   *
-   * @param bytesToBeAvailable the requested bytes to be available
-   * @param plan the eviction plan, should not be null
-   * @param metaManager the meta data manager
-   * @throws Exception when fail
-   */
-  public static void assertEvictionPlanValid(long bytesToBeAvailable, EvictionPlan plan,
-      BlockMetadataManager metaManager) throws Exception {
-    Assert.assertNotNull(plan);
-    Assert.assertTrue(validNonCascadingPlan(bytesToBeAvailable, plan, metaManager)
-        || TieredBlockStoreTestUtils.validCascadingPlan(bytesToBeAvailable, plan, metaManager));
-  }
 }
